backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
Faker==33.0.0
flake8==7.1.1
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
pep8-naming==0.14.1
pillow==11.0.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
